Fixed complex entry point
